+++ src/lib.rs
@@ -28,11 +28,7 @@
 #[doc = include_str!("../README.md")]
 mod readme {}
 
-mod assert;
 mod format;
-mod utils;
-mod week;
-mod write;
 
 #[cfg(test)]
 mod tests;
